+++ TP1_Integrador/src/Entities/Producto.java
@@ -3,10 +3,10 @@ package Entities;
 import java.util.Objects;
 
 public class Producto {
-	private int idProducto;
+	private Long idProducto;
 	private String nombre;
 	private float valor;
-	public Producto(int idProducto, String nombre, float valor) {
+	public Producto(Long idProducto, String nombre, float valor) {
 		super();
 		this.idProducto = idProducto;
 		this.nombre = nombre;
